refactor(ui): align imports to shared lib
